+++ examples/test-partial/inmanta_plugins/test_partial/__init__.py
@@ -1,0 +1,7 @@
+"""
+    Pytest Inmanta LSM
+
+    :copyright: 2022 Inmanta
+    :contact: code@inmanta.com
+    :license: Inmanta EULA
+"""