+++ src/pytest_inmanta_lsm/resources/setup_project.py
@@ -15,11 +15,13 @@
 
     Contact: code@inmanta.com
 """
+import contextlib
 import logging
 import os
 import pathlib
 import sys
-from typing import List
+from collections import abc
+from typing import List, Optional
 
 from inmanta import env, module
 
@@ -43,6 +45,24 @@ except Exception as e:
 logging.root.handlers = []
 logging.root.addHandler(stream_handler)
 logging.root.setLevel(logging.DEBUG)
+
+
+@contextlib.contextmanager
+def env_vars(var: abc.Mapping[str, str]) -> abc.Iterator[None]:
+    """
+    Context manager to extend the current environment with one or more environment variables.
+    """
+
+    def set_env(set_var: abc.Mapping[str, Optional[str]]) -> None:
+        for name, value in set_var.items():
+            if value is not None:
+                os.environ[name] = value
+
+    old_env: abc.Mapping = {name: os.environ.get(name, None) for name in var}
+    set_env(var)
+    yield
+    set_env(old_env)
+
 
 # Create the project object, this is the folder we sent to the orchestrator
 project = module.Project(str(project_path), venv_path=str(project_path / ".env"))
@@ -76,10 +96,15 @@ for dir in (project_path / "libs").iterdir():
     v2_modules.append(mod)
     LOGGER.info(f"Module {mod.name} is v2, we will attempt to install it")
 
-# Install all v2 modules in editable mode
+# Install all v2 modules in editable mode using the project's configured package sources
 if v2_modules:
-    LOGGER.info(f"Installing modules from source: {[mod.name for mod in v2_modules]}")
-    project.virtualenv.install_from_source([env.LocalPackagePath(mod.path, editable=True) for mod in v2_modules])
+    urls: abc.Sequence[str] = project.module_source.urls
+    if not urls:
+        raise Exception("No package repos configured for project")
+    # plain Python install so core does not apply project's sources -> we need to configure pip index ourselves
+    with env_vars({"PIP_INDEX_URL": urls[0], "PIP_EXTRA_INDEX_URL": " ".join(urls[1:])}):
+        LOGGER.info(f"Installing modules from source: {[mod.name for mod in v2_modules]}")
+        project.virtualenv.install_from_source([env.LocalPackagePath(mod.path, editable=True) for mod in v2_modules])
 
 # Install all other dependencies
 LOGGER.info("Installing other project dependencies")